Fixed counter problem Added new generic test module All test passed at this point

+++ ECGScanner/globals.h
@@ -1,7 +1,7 @@
 #ifndef GLOBALS_H
 #define GLOBALS_H
 
-extern int GLOBAL_COUNT;
+extern unsigned long int GLOBAL_COUNT;
 extern int GLOBAL_SAMPLE_RATE;
 extern int GLOBAL_SENSOR_ACTIVE;
 extern char GLOBAL_SENSOR_INPUT[];
